fix(addon): Restore "unknown" string for empty 'user' field
This is not in-line with Splunk best practice which recommends to have
the field be null instead of contain a string, but we need to have
something here for the logs to show up in the dashboards. We would have
to adjust all dashboards to remove the "unknown" string. This will be
done at a later date, but for now, we need to put back the "unknown".