Fix common.py get_firewall_credentials
common.py was throwing several errors.  The first was a local variable 'password' referenced before assignment.  To resolve this, I defined the password variable before the get_firewall_credentials and added the global password statement.

Once this was resolved, I received the message that no user or password were defined for the searchcommand.  This wasn't accurate, but I found that the function was looking for

'Firewall``splunk_cred_sep``1'

but should have been looking for:

'firewall``splunk_cred_sep``1'